requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
django-cors-headers==4.1.0
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.0
Pillow==9.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==0.21.1
pytz==2025.1
PyYAML==6.0.1
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
whitenoise==6.5.0
